Another mid-run commit of mpi_xy8.f90
I'm in the process of running the code now (23 May '19) and the code is
producing results that are nearly identical to the ones for the original
code (albeit with different error bars due to the Monte Carlo sim, but
c'est la vie.)

TODO:
1) Run another one with 30% impurity in the system (as opposed to a clean
run, which is the standard basis of comparison.)
2) Clean up the code, and remove the erroneous stuf that pertains to
the quantum implementation (i.e. the time dependent/time axis stuff),
and make a single length parameter that sets the system size.
3) With that, add command line flags, so that the code can be compiled
somewhat more dynamically than having to go in and change numbers every
time you want a different system size.
4) ...

Another commit will probably come after the code's done running. -TVH