Added new test.f90 file/minor cleanup to mpi_xy8
test.f90 is an experiment in trying to implement command line arguments
for the main file so there's not so much clunky hardcoding. The minor
cleanups to mpi_xy8 are just removing some unused variables and fixing
indenting so that the fortran compiler doesn't have a fit.

More to follow. -TVH