New changes to mpi_xy8/_rework
Per Dr Vojta's request, I modified the mpi_xy8 code so that there's a
single length parameter instead of a length parameter and a time
parameter. It's running at the moment, but I like the numbers I'm
seeing. mpi_xy8_rework is where those modifications were supposed to go,
but the gfortran compiler threw a fit and wouldn't execute it.

More to come. -TVH