Big update - should have done this earlier
Well, I just about royally screwed myself. Like, really royally.
gfortran did something absolutely insane when I compiled my file, and
everything was nearly lost (at least, from the last commit, which was a
long-ass time ago). This code in the mpi_xy8 file is now perfect. This
is the standard that should be used for the classical model.

More to come. - TVH